Fixed Profit

+++ MedicalProject/Models/Order/OrderDto.cs
@@ -14,7 +14,7 @@
         //قیمت هر دانگ
         public string PricePerDong { get; set; }
         //مقدار خواسته شده
-        public int DongAmount { get; set; }
+        public decimal DongAmount { get; set; }
         public Guid InventoryId { get; set; }
 
         public decimal TotalPrice
